feat: 폴더 정리 및 디렉토리 수정/삭제 모달 구현
