+++ pages/_document.tsx
@@ -1,8 +1,6 @@
 import { Head, Html, Main, NextScript } from 'next/document';
-import { useRef } from 'react';
 
 export default function Document() {
-
   return (
     <Html lang="ko">
       <Head>
@@ -19,9 +17,9 @@ export default function Document() {
         />
       </Head>
       <body>
+      <div id="modal-root"></div>
         <Main />
         <NextScript />
-        <div id="modal-root"></div>
       </body>
     </Html>
   );
